back/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
filetype==1.2.0
Flask==3.1.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.49
langchain-google-genai==2.1.2
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.19
MarkupSafe==3.0.2
openai==1.69.0
orjson==3.10.16
packaging==24.2
priority==2.0.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
